move mount health to own mixin
