+++ src/main/resources/thermoo.client.mixins.json
@@ -4,6 +4,7 @@
     "compatibilityLevel": "JAVA_17",
     "plugin": "com.github.thedeathlycow.thermoo.mixin.client.compat.Plugin",
     "client": [
+        "InGameHudMixin",
         "compat.colorfulhearts.present.HeartRendererMixin",
         "compat.overflowingbars.absent.InGameHudMixin",
         "compat.overflowingbars.present.HealthBarRendererMixin"
